refactor(attempt-review): split review page components
