+++ src/course/attempt-review/attempt-review-comment-annotation.model.ts
@@ -1,0 +1,5 @@
+import type { AttemptReviewLineComment } from './attempt-review.types';
+
+export interface AttemptReviewLineCommentAnnotation {
+  comment: AttemptReviewLineComment;
+}
