+++ src/course/attempt-review/attempt-review-comment-timestamp.component.tsx
@@ -1,0 +1,23 @@
+import { formatAttemptReviewRelativeDateTime } from './attempt-review-date.format';
+
+interface AttemptReviewCommentTimestampProps {
+  createdAt: string;
+  updatedAt: string;
+}
+
+export function AttemptReviewCommentTimestamp({
+  createdAt,
+  updatedAt,
+}: AttemptReviewCommentTimestampProps) {
+  const edited = createdAt !== updatedAt;
+  const date = formatAttemptReviewRelativeDateTime(
+    edited ? updatedAt : createdAt
+  );
+
+  return (
+    <span title={date.title}>
+      {date.label}
+      {edited ? ' · изменено' : ''}
+    </span>
+  );
+}
