+++ src/course/attempt-review/attempt-review-diff.component.tsx
@@ -1,19 +1,35 @@
 import { useCallback, useMemo, useRef, useState, type RefObject } from 'react';
-import type { CodeViewItem, CodeViewOptions } from '@pierre/diffs';
+import type { CodeViewOptions } from '@pierre/diffs';
 import { CodeView, type CodeViewHandle } from '@pierre/diffs/react';
-import { ArrowUp, ChevronDown, ChevronRight } from 'lucide-react';
+import { ArrowUp } from 'lucide-react';
 
+import { AttemptReviewCodeViewHeader } from './attempt-review-code-view-header.component';
 import {
-  getAttemptReviewFileStatusGlyph,
-  getAttemptReviewFileStatusIconClassName,
-} from './attempt-review-file-status.format';
+  createAttemptReviewCodeViewItems,
+  groupAttemptReviewCommentsByFile,
+} from './attempt-review-code-view-item.model';
+import type { AttemptReviewLineCommentAnnotation } from './attempt-review-comment-annotation.model';
+import {
+  addLineComment,
+  cancelLineComment,
+  deleteLineComment,
+  deleteLineCommentReply,
+  editLineComment,
+  revertPendingLineComment,
+  submitLineComment,
+  submitLineCommentReply,
+  updateLineCommentReply,
+} from './attempt-review-line-comment.model';
+import {
+  canManageLineComment,
+  canReplyToLineComment,
+} from './attempt-review-line-comment-permissions.model';
 import { useAttemptReviewScrollHandoff } from './attempt-review-scroll-handoff.hook';
 import { useHtmlThemeType } from './attempt-review-theme';
 import { AttemptReviewLineCommentCard } from './attempt-review-line-comment-card.component';
 import type {
   AttemptReviewChangedFile,
   AttemptReviewCommentAuthor,
-  AttemptReviewCommentSide,
   AttemptReviewLineComment,
   AttemptReviewMode,
 } from './attempt-review.types';
@@ -38,10 +54,6 @@ interface AttemptReviewDiffProps {
   onViewerChange?: (
     viewer: CodeViewHandle<AttemptReviewLineCommentAnnotation> | null
   ) => void;
-}
-
-export interface AttemptReviewLineCommentAnnotation {
-  comment: AttemptReviewLineComment;
 }
 
 const CODE_VIEW_LAYOUT = {
@@ -92,7 +104,7 @@ export function AttemptReviewDiff({
     () => new Set()
   );
   const commentsByFile = useMemo(
-    () => groupCommentsByFile(comments),
+    () => groupAttemptReviewCommentsByFile(comments),
     [comments]
   );
   const filesByPath = useMemo(
@@ -110,24 +122,12 @@ export function AttemptReviewDiff({
     innerScrollRef: codeViewContainerRef,
   });
 
-  const items = useMemo<CodeViewItem<AttemptReviewLineCommentAnnotation>[]>(
+  const items = useMemo(
     () =>
-      files.map((file) => {
-        const fileComments = commentsByFile.get(file.path) ?? [];
-        const collapsed = collapsedFilePaths.has(file.path);
-
-        return {
-          id: file.path,
-          type: 'diff',
-          fileDiff: file.diff,
-          collapsed,
-          version: buildItemVersion(file.path, collapsed, fileComments),
-          annotations: fileComments.map((comment) => ({
-            side: comment.endSide ?? comment.side,
-            lineNumber: comment.endLineNumber ?? comment.lineNumber,
-            metadata: { comment },
-          })),
-        };
+      createAttemptReviewCodeViewItems({
+        files,
+        commentsByFile,
+        collapsedFilePaths,
       }),
     [collapsedFilePaths, commentsByFile, files]
   );
@@ -151,13 +151,14 @@ export function AttemptReviewDiff({
                 return;
               }
 
-              addLineComment({
-                comments,
-                filePath: context.item.id,
-                range,
-                currentReviewer,
-                onCommentsChange,
-              });
+              onCommentsChange?.(
+                addLineComment({
+                  comments,
+                  filePath: context.item.id,
+                  range,
+                  currentReviewer,
+                })
+              );
             }
           : undefined,
     }),
@@ -192,68 +193,33 @@ export function AttemptReviewDiff({
     codeViewHandleRef.current?.clearSelectedLines();
   }
 
-  function updateComment(
-    commentId: string,
-    update: (
-      comment: AttemptReviewLineComment
-    ) => AttemptReviewLineComment | null
-  ): AttemptReviewLineComment[] {
-    const nextComments = comments.flatMap((comment) => {
-      if (comment.id !== commentId) {
-        return [comment];
-      }
-
-      const nextComment = update(comment);
-      return nextComment ? [nextComment] : [];
-    });
-
+  function applyComments(nextComments: AttemptReviewLineComment[]) {
     onCommentsChange?.(nextComments);
     return nextComments;
   }
 
   function submitComment(commentId: string, html: string) {
-    updateComment(commentId, (comment) => ({
-      ...comment,
-      html,
-      status: comment.status === 'draft' ? 'pending-create' : 'pending-update',
-      isEditing: false,
-    }));
+    applyComments(submitLineComment(comments, commentId, html));
     clearSelectedLines();
   }
 
   function cancelComment(commentId: string) {
-    updateComment(commentId, (comment) => {
-      if (comment.status === 'draft') {
-        return null;
-      }
-
-      return { ...comment, isEditing: false };
-    });
+    applyComments(cancelLineComment(comments, commentId));
     clearSelectedLines();
   }
 
   function editComment(commentId: string) {
-    updateComment(commentId, (comment) => ({ ...comment, isEditing: true }));
+    applyComments(editLineComment(comments, commentId));
   }
 
   function deleteComment(commentId: string) {
-    updateComment(commentId, (comment) => {
-      if (comment.status === 'draft' || comment.status === 'pending-create') {
-        return null;
-      }
-
-      return { ...comment, status: 'pending-delete', isEditing: false };
-    });
+    applyComments(deleteLineComment(comments, commentId));
   }
 
   function revertPendingComment(commentId: string) {
-    updateComment(commentId, (comment) => {
-      if (comment.status === 'pending-create') {
-        return null;
-      }
-
-      return savedCommentsById.get(comment.id) ?? null;
-    });
+    applyComments(
+      revertPendingLineComment(comments, commentId, savedCommentsById)
+    );
   }
 
   async function submitReply(commentId: string, html: string) {
@@ -261,42 +227,30 @@ export function AttemptReviewDiff({
       return;
     }
 
-    const now = new Date().toISOString();
-    const nextComments = updateComment(commentId, (comment) => ({
-      ...comment,
-      replies: [
-        ...(comment.replies ?? []),
-        {
-          id: `reply-${comment.id}-${Date.now()}`,
-          html,
-          authorName: currentReviewer.name,
-          authorUsername: currentReviewer.username,
-          createdAt: now,
-          updatedAt: now,
-        },
-      ],
-    }));
+    const nextComments = applyComments(
+      submitLineCommentReply({
+        comments,
+        commentId,
+        html,
+        currentReviewer,
+      })
+    );
 
     await onCommentsPersist?.(nextComments);
   }
 
   async function updateReply(commentId: string, replyId: string, html: string) {
-    const now = new Date().toISOString();
-    const nextComments = updateComment(commentId, (comment) => ({
-      ...comment,
-      replies: (comment.replies ?? []).map((reply) =>
-        reply.id === replyId ? { ...reply, html, updatedAt: now } : reply
-      ),
-    }));
+    const nextComments = applyComments(
+      updateLineCommentReply({ comments, commentId, replyId, html })
+    );
 
     await onCommentsPersist?.(nextComments);
   }
 
   async function deleteReply(commentId: string, replyId: string) {
-    const nextComments = updateComment(commentId, (comment) => ({
-      ...comment,
-      replies: (comment.replies ?? []).filter((reply) => reply.id !== replyId),
-    }));
+    const nextComments = applyComments(
+      deleteLineCommentReply({ comments, commentId, replyId })
+    );
 
     await onCommentsPersist?.(nextComments);
   }
@@ -335,7 +289,7 @@ export function AttemptReviewDiff({
           }
 
           return (
-            <CodeViewFileHeader
+            <AttemptReviewCodeViewHeader
               file={file}
               collapsed={item.collapsed === true}
               onToggleCollapsed={() => toggleFileCollapsed(item.id)}
@@ -391,174 +345,4 @@ export function AttemptReviewDiff({
       ) : null}
     </div>
   );
-}
-
-function canManageLineComment(comment: AttemptReviewLineComment): boolean {
-  return (comment.status ?? 'saved') === 'saved';
-}
-
-function canReplyToLineComment(comment: AttemptReviewLineComment): boolean {
-  return (comment.status ?? 'saved') === 'saved';
-}
-
-function CodeViewFileHeader({
-  file,
-  collapsed,
-  onToggleCollapsed,
-}: {
-  file: AttemptReviewChangedFile;
-  collapsed: boolean;
-  onToggleCollapsed: () => void;
-}) {
-  return (
-    <div className="flex h-14 min-w-0 items-center justify-between gap-3 border-b bg-card/95 px-4 text-sm backdrop-blur supports-[backdrop-filter]:bg-card/85">
-      <div className="flex min-w-0 items-center gap-2 font-medium text-card-foreground">
-        <button
-          type="button"
-          className="grid size-7 shrink-0 cursor-pointer place-items-center rounded-md text-muted-foreground transition-colors hover:bg-muted hover:text-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-          aria-label={
-            collapsed ? 'Показать изменение файла' : 'Скрыть изменение файла'
-          }
-          title={
-            collapsed ? 'Показать изменение файла' : 'Скрыть изменение файла'
-          }
-          aria-expanded={!collapsed}
-          onClick={onToggleCollapsed}
-        >
-          {collapsed ? (
-            <ChevronRight className="size-4" />
-          ) : (
-            <ChevronDown className="size-4" />
-          )}
-        </button>
-        <span
-          className={`flex size-5 shrink-0 items-center justify-center rounded-md border text-xs leading-none ${getAttemptReviewFileStatusIconClassName(file.status)}`}
-          aria-hidden="true"
-        >
-          <span className="-translate-y-px leading-none">
-            {getAttemptReviewFileStatusGlyph(file.status)}
-          </span>
-        </span>
-        <span className="min-w-0 truncate">{file.path}</span>
-      </div>
-      <div className="flex shrink-0 items-center gap-2 text-xs font-semibold">
-        {file.deletedLines > 0 || file.addedLines === 0 ? (
-          <span className="text-rose-500">−{file.deletedLines}</span>
-        ) : null}
-        {file.addedLines > 0 || file.deletedLines === 0 ? (
-          <span className="text-emerald-500">+{file.addedLines}</span>
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
-function buildItemVersion(
-  filePath: string,
-  collapsed: boolean,
-  comments: AttemptReviewLineComment[]
-): number {
-  return hashText(
-    JSON.stringify({
-      filePath,
-      collapsed,
-      comments: comments.map((comment) => ({
-        id: comment.id,
-        side: comment.side,
-        lineNumber: comment.lineNumber,
-        endSide: comment.endSide,
-        endLineNumber: comment.endLineNumber,
-        html: comment.html,
-        authorUsername: comment.authorUsername,
-        createdAt: comment.createdAt,
-        updatedAt: comment.updatedAt,
-        status: comment.status,
-        isEditing: comment.isEditing,
-        replies: comment.replies,
-      })),
-    })
-  );
-}
-
-function hashText(value: string): number {
-  let hash = 0;
-
-  for (let index = 0; index < value.length; index += 1) {
-    hash = (hash * 31 + value.charCodeAt(index)) | 0;
-  }
-
-  return hash;
-}
-
-function groupCommentsByFile(comments: AttemptReviewLineComment[]) {
-  const grouped = new Map<string, AttemptReviewLineComment[]>();
-
-  comments.forEach((comment) => {
-    grouped.set(comment.filePath, [
-      ...(grouped.get(comment.filePath) ?? []),
-      comment,
-    ]);
-  });
-
-  return grouped;
-}
-
-function addLineComment({
-  comments,
-  filePath,
-  range,
-  currentReviewer,
-  onCommentsChange,
-}: {
-  comments: AttemptReviewLineComment[];
-  filePath: string;
-  currentReviewer: AttemptReviewCommentAuthor;
-  range: {
-    start: number;
-    side?: AttemptReviewCommentSide;
-    end: number;
-    endSide?: AttemptReviewCommentSide;
-  };
-  onCommentsChange: AttemptReviewDiffProps['onCommentsChange'];
-}) {
-  onCommentsChange?.([
-    ...comments,
-    createDraftLineComment({ filePath, range, currentReviewer }),
-  ]);
-}
-
-function createDraftLineComment({
-  filePath,
-  range,
-  currentReviewer,
-}: {
-  filePath: string;
-  range: {
-    start: number;
-    side?: AttemptReviewCommentSide;
-    end: number;
-    endSide?: AttemptReviewCommentSide;
-  };
-  currentReviewer: AttemptReviewCommentAuthor;
-}): AttemptReviewLineComment {
-  const side = range.side ?? range.endSide ?? 'additions';
-  const endSide = range.endSide ?? side;
-  const lineNumber = range.start;
-  const endLineNumber = range.end;
-
-  return {
-    id: `draft-${filePath}-${side}-${lineNumber}-${endSide}-${endLineNumber}-${Date.now()}`,
-    filePath,
-    side,
-    lineNumber,
-    endSide:
-      endSide === side && endLineNumber === lineNumber ? undefined : endSide,
-    endLineNumber: endLineNumber === lineNumber ? undefined : endLineNumber,
-    html: '<p></p>',
-    authorName: currentReviewer.name,
-    authorUsername: currentReviewer.username,
-    createdAt: '',
-    updatedAt: '',
-    status: 'draft',
-  };
 }
