+++ src/course/attempt-review/attempt-review-line-comment-permissions.model.ts
@@ -1,0 +1,13 @@
+import type { AttemptReviewLineComment } from './attempt-review.types';
+
+export function canManageLineComment(
+  comment: AttemptReviewLineComment
+): boolean {
+  return (comment.status ?? 'saved') === 'saved';
+}
+
+export function canReplyToLineComment(
+  comment: AttemptReviewLineComment
+): boolean {
+  return (comment.status ?? 'saved') === 'saved';
+}
